# Conflicts:
#   Siesa.SDK.Frontend/Components/FormManager/Views/DetailView.razor.cs

+++ Siesa.SDK.Frontend/Components/FormManager/Views/DetailView.razor.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
@@ -32,11 +32,11 @@ namespace Siesa.SDK.Frontend.Components.FormManager.Views
         public bool ShowTitle { get; set; } = true;
         [Parameter]
         public bool ShowButtons { get; set; } = true;
-        [Parameter] 
-        public bool ShowCancelButton {get; set;} = true;
-                
-        [Parameter] 
-        public bool ShowDeleteButton {get; set;} = true;
+        [Parameter]
+        public bool ShowCancelButton { get; set; } = true;
+
+        [Parameter]
+        public bool ShowDeleteButton { get; set; } = true;
 
         [Inject] public IJSRuntime JSRuntime { get; set; }
         [Inject] public NavigationManager NavManager { get; set; }
@@ -157,7 +157,7 @@ namespace Siesa.SDK.Frontend.Components.FormManager.Views
         {
             var result = await BusinessObj.DeleteAsync();
 
-            if (result.Errors.Count > 0)
+            if (result != null && result.Errors.Count > 0)
             {
                 ErrorMsg = "<ul>";
                 foreach (var error in result.Errors)
@@ -170,18 +170,21 @@ namespace Siesa.SDK.Frontend.Components.FormManager.Views
                 ErrorMsg += "</ul>";
                 return;
             }
-            if (IsSubpanel)
+            if (result != null)
             {
-                dialogService.Close(false);
-            }
-            else
-            {                
-                string uri = NavManager.Uri;
-                NavigationService.RemoveItem(uri);
-                NavManager.NavigateTo($"{BusinessName}/");
+                if (IsSubpanel)
+                {
+                    dialogService.Close(false);
+                }
+                else
+                {
+                    string uri = NavManager.Uri;
+                    NavigationService.RemoveItem(uri);
+                    NavManager.NavigateTo($"{BusinessName}/");
+                }
             }
         }
-        
+
 
         private void OnClickCustomButton(Button button)
         {
